// src/types/Monitoring.types.d.ts
export interface thresholdType {
  vibration?: {
    red?: number;
    yellow?: number;
    green?: number;
  };
  noise?: {
    red?: number;
    yellow?: number;
    green?: number;
  };
  temperature?: {
    red?: number;
    yellow?: number;
    green?: number;
  };
}
// ...
export interface deviceHistoryType {
  id: number;
  time: string;
  value: string;
  status: string;
  key: string;
}
// ...
export interface historyChartProps {
  chartData: deviceHistoryType[];
  deviceThreshold: thresholdType;
  metric: string;
}

export interface chartVertexType {
  peakValue: number;
  averageValue: number;
  goodValue: number;
  presentValue: number;
  time: string;
}